# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Hospital Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenInput"
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin)
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Deactivate/delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /users/{id}/reset-password:
    post:
      operationId: adminResetPassword
      tags: [users]
      summary: Admin resets a user password
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminResetPasswordInput"
      responses:
        "200":
          description: Password reset

  # ── PATIENTS ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    RefreshTokenInput:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    AuthResponse:
      type: object
      required: [accessToken, refreshToken, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/User"
...
    AdminResetPasswordInput:
      type: object
      required: [newPassword]
      properties:
        newPassword:
          type: string

    # ── USER SCHEMAS ─────────────────────────────────────────────────────────
    User:
      type: object
      required: [id, username, role, fullName, isActive, createdAt]
      properties:
        id:
          type: string
        username:
          type: string
        role:
          type: string
          enum: [admin, doctor, staff]
        fullName:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        registrationNumber:
          type: ["string", "null"]
        specialization:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    UserInput:
      type: object
      required: [username, password, role, fullName]
      properties:
        username:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, doctor, staff]
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        registrationNumber:
          type: string
        specialization:
          type: string
    UserUpdate:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        registrationNumber:
          type: string
        specialization:
          type: string
        role:
          type: string
          enum: [admin, doctor, staff]
        isActive:
          type: boolean
    UserListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ── PATIENT SCHEMAS ──────────────────────────────────────────────────────
...